Add a new md_add_new_row_to_sorted_list function to insert a row into a list column's target table in the correctly sorted position so we do not need to re-sort later.
This functionality will be used in the IMetaDataEmit implementation to ensure that new Param table entries are specified in Sequence order.

This is something that the runtime depends on (and DNMD already validates in the md_apply_delta function), but is not specified in ECMA-335 or in the ECMA-335 Augments doc. As a result, I don't think we need to block developers from using md_add_new_row_to_list with the ParamList column (we also use this function in the apply-delta functionality for this case). Instead, we can rely on users to use the API that provides the guarantees necessary for their use case.